+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All new features, changes and fixes should be listed here. Please use tickets to reference changes.
 
-## 1.0.0 (2014/xx/xx)
+## 1.0.0 (2014/07/01)
 
 The `Parameters` class now extends `ArrayObject`. That means, in addtion to
 `ArrayAccess` it implements `Serializable`, `Countable` and `IteratorAggregate`.